misc: package the devtool plugins
eb6a14882017db249ef7cff8a8cfff13e92d2151 has split the devtool plugins into a separate component,
which caused them not to be packaged as part of the existing .deb, however this can still be useful
to have them.

This commit adds the new deb vpp-plugin-devtools which contains that component.

Change-Id: I3cf44493745c3d4951ffd2194c6ae539e8ad5926
Type: fix
Signed-off-by: Andrew Yourtchenko <ayourtch@gmail.com>